+++ tests/rewrite/fixtures/sample_helper_find_device_mutated_image_max.cu
@@ -1,0 +1,7 @@
+#include <helper_string.h>
+#include <helper_functions.h>
+#include <helper_cuda.h>
+
+int rejectMutatedImageMaxFindDevice(int argc, const char** argv) {
+  return findCudaDevice(argc, argv);
+}